Split spectator overlay into separate file
